+++ src/main/java/com/qwqnoi/community/controller/QuestionController.java
@@ -15,7 +15,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable Integer id, Model model) {
+    public String question(@PathVariable Long id, Model model) {
         QuestionDTO questionDTO = questionService.getById(id);
         //累加评论
         questionService.incView(id);
